pyproject.toml
[tool.poetry]
name = "fastapi-middleware"
version = "0.1.0"
description = "A FastAPI middleware application"
authors = ["Cozyamy <cozyamy@gmail.com>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
fastapi = "^0.110.1"
uvicorn = "^0.29.0"
pydantic = "^2.6.4"
pytest = "^8.1.1"
python-magic = "^0.4.27"
pytest-cov = "^5.0.0"
httpx = "^0.27.0"
python-multipart = "^0.0.9"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
